Dataset
Empowering Farmers with Soil Insights (used in crop selection):
DataCamp published a soil_measures.csv dataset to use.
Each row in this dataset represents various measures of the soil in a particular field. Based on these measurements, the crop specified in the "crop" column is the optimal choice for that field.
This data can be used to predict the best crop for a farmer's field.